backend/requirements.txt
# Core
fastapi
uvicorn
python-dotenv

# Database
sqlalchemy
psycopg2-binary

# Vector Database
pinecone

# AI/LLM
groq
langgraph
langchain
langchain-groq

# Validation
pydantic
email-validator

# Utilities
httpx
